refactoting for better modularity, still not functional, TODO: animation
