+++ .gitignore
@@ -20,7 +20,7 @@
 .idea/Proyecto-de-Grado-Codes.iml
 matriz_u.txt
 busy_widget.py
-/simulation/build
+/app/simulation/build
 laplacian.pyd
 *.pyc
 laplacian.html
